test(plugin-solid): split Babel transform order fixture
