+++ e2e/cases/plugin-solid/babel-transform-order/src/index.jsx
@@ -1,0 +1,5 @@
+import { render } from '@solidjs/web';
+
+const App = () => <button id="button">button</button>;
+
+render(App, document.getElementById('root'));
